crawler/cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "regulaai-cli"
version = "0.1.0"
description = "RegulaAI CLI - GDPR Compliance Scanning Tool"
authors = [
    { name = "RegulaAI", email = "support@regulaai.com" }
]
readme = "README.md"
license = { file = "../LICENSE" }
requires-python = ">=3.8"
dependencies = [
    "typer[all] >=0.9.0",
    "requests >=2.31.0",
    "rich >=13.0.0",
    "pyyaml >=6.0.0",
    "sqlalchemy >=2.0.0"
]

[project.scripts]
regulaai = "cli.main:app"

[tool.setuptools]
packages = ["cli"]
include-package-data = true 